+++ src/main/java/com/sparta/delivery/review/controller/ReviewController.java
@@ -1,13 +1,9 @@
-package com.sparta.delivery.review;
+package com.sparta.delivery.review.controller;
 
 import com.sparta.delivery.common.ResponseCode;
 import com.sparta.delivery.common.dto.ResponseDto;
 import com.sparta.delivery.common.dto.ResponsePageDto;
-import com.sparta.delivery.common.dto.ResponseSingleDto;
-import com.sparta.delivery.product.dto.ProductCreateRequestDto;
-import com.sparta.delivery.product.dto.ProductListResponseDto;
-import com.sparta.delivery.product.dto.ProductModifyRequestDto;
-import com.sparta.delivery.product.dto.ProductResponseDto;
+import com.sparta.delivery.review.service.ReviewService;
 import com.sparta.delivery.review.dto.ReviewCreateRequestDto;
 import com.sparta.delivery.review.dto.ReviewListResponseDto;
 import com.sparta.delivery.review.dto.ReviewModifyRequestDto;
@@ -16,6 +12,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,34 +28,50 @@ public class ReviewController {
 
     /**
      * 리뷰 생성
+     * @param dto           리뷰 정보 dto
+     * @param principal     사용자 정보를 담고 있는 객체
      */
     @PreAuthorize("hasRole('CUSTOMER')")
     @PostMapping
-    public ResponseDto createReview(@RequestBody ReviewCreateRequestDto dto, Principal principal) {
-        reviewService.createReview(dto,principal);
-        return ResponseDto.of(ResponseCode.SUCC_REVIEW_CREAET);
+    public ResponseEntity createReview(@RequestBody ReviewCreateRequestDto dto, Principal principal) {
+        reviewService.createReview(dto, principal);
+
+        ResponseDto response = ResponseDto.of(ResponseCode.SUCC_REVIEW_CREAET);
+        return ResponseEntity.ok(response);
     }
 
     /**
      * 리뷰 사용자 목록조회
+     * @param storeId   음식점 식별자
+     * @param page      조회 페이지
+     * @param size      조회 페이지 사이즈
      */
     @PreAuthorize("hasRole('CUSTOMER')")
     @GetMapping("/{storeId}")
-    public ResponsePageDto getReviewList(@PathVariable UUID storeId,
+    public ResponseEntity getReviewList(@PathVariable UUID storeId,
                                          @RequestParam(required = false, defaultValue = "1") int page,
                                           @RequestParam(required = false, defaultValue = "10") int size) {
 
         Pageable pageable=PageRequest.of(page - 1, size, Sort.by("createdAt").descending());
         Page<ReviewListResponseDto> data = reviewService.getReviewList(storeId,pageable);
-        return ResponsePageDto.of(ResponseCode.SUCC_REVIEW_USER_LIST_GET, data);
+
+        ResponsePageDto response = ResponsePageDto.of(ResponseCode.SUCC_REVIEW_USER_LIST_GET, data);
+        return ResponseEntity.ok(response);
     }
 
     /**
      * 리뷰 사장님 목록조회
+     * @param storeId   음식점 식별자
+     * @param keyWord   검색어
+     * @param type      검색어의 타입
+     * @param page      조회 페이지
+     * @param size      조회 페이지 사이즈
+     * @param sort      정렬 기준
+     * @param asc       정렬 방향
      */
     @PreAuthorize("hasRole('OWNER') and @securityUtil.isStoreOwner(authentication,#storeId) ")
     @GetMapping("/owner/{storeId}")
-    public ResponsePageDto getOwnerReviewList(@PathVariable UUID storeId,
+    public ResponseEntity getOwnerReviewList(@PathVariable UUID storeId,
                                               @RequestParam(required = false, defaultValue = "") String keyWord,
                                               @RequestParam(required = false, defaultValue = "content") String type,
                                               @RequestParam(required = false, defaultValue = "1") int page,
@@ -69,42 +82,55 @@ public class ReviewController {
         Pageable pageable = asc ? PageRequest.of(page - 1, size, Sort.by(sort).ascending()) :
                 PageRequest.of(page - 1, size, Sort.by(sort).descending());
         Page<ReviewListResponseDto> data = reviewService.getOwnerReviewList(storeId, keyWord, type, pageable);
-        return ResponsePageDto.of(ResponseCode.SUCC_REVIEW_LIST_GET, data);
+
+        ResponsePageDto response = ResponsePageDto.of(ResponseCode.SUCC_REVIEW_LIST_GET, data);
+        return ResponseEntity.ok(response);
     }
 
     /**
      * 리뷰 수정
-     * 해당 리뷰의 생성자일 경우
+     * @param reviewId  리뷰 식별자
+     * @param dto       리뷰 수정 정보 dto
+     * 리소스 접근 사용자와  리뷰 생성자가 동일한지 검사
      */
     @PreAuthorize("hasRole('CUSTOMER') and @securityUtil.isReivewOwner(authentication,#reviewId)")
     @PutMapping("/{reviewId}")
-    public ResponseDto ModifyReview(@PathVariable UUID reviewId,
+    public ResponseEntity ModifyReview(@PathVariable UUID reviewId,
                                     @RequestBody ReviewModifyRequestDto dto) {
-        reviewService.modifyProduct(reviewId, dto);
-        return ResponseDto.of(ResponseCode.SUCC_REVIEW_MODIFY);
+        reviewService.modifyreview(reviewId, dto);
+
+        ResponseDto response = ResponseDto.of(ResponseCode.SUCC_REVIEW_MODIFY);
+        return ResponseEntity.ok(response);
     }
 
     /**
      * 리뷰 삭제
-     * 해당 리뷰의 생성자일 경우
+     * @param reviewId  리뷰 식별자
+     * 리소스 접근 사용자와  리뷰 생성자가 동일한지 검사
      */
     @PreAuthorize("hasRole('CUSTOMER') and @securityUtil.isReivewOwner(authentication,#reviewId)")
     @DeleteMapping("/{reviewId}")
-    public ResponseDto deleteReview(@PathVariable UUID reviewId) {
-        reviewService.deleteProduct(reviewId);
-        return ResponseDto.of(ResponseCode.SUCC_REVIEW_DELETE);
+    public ResponseEntity deleteReview(@PathVariable UUID reviewId) {
+        reviewService.deleteReview(reviewId);
+
+        ResponseDto response = ResponseDto.of(ResponseCode.SUCC_REVIEW_DELETE);
+        return ResponseEntity.ok(response);
     }
 
     /**
      * 리뷰 신고
+     * @param reviewId  리뷰 식별자
+     * @param content   리뷰 내용
      */
-    //content를 body로? 쿼리 스트링으로? -> dto 유효성 검사
+    //todo : content를 body로? 쿼리 스트링으로? -> dto 유효성 검사
     @PreAuthorize("hasRole('CUSTOMER')")
     @PatchMapping("/{reviewId}")
-    public ResponseDto ReportReview(@PathVariable UUID reviewId,
-                                    @RequestParam(required = true) String content) {
+    public ResponseEntity ReportReview(@PathVariable UUID reviewId,
+                                    @RequestParam String content) {
         reviewService.ReportReview(reviewId,content);
-        return ResponseDto.of(ResponseCode.SUCC_REVIEW_REPORT);
+
+        ResponseDto response = ResponseDto.of(ResponseCode.SUCC_REVIEW_REPORT);
+        return ResponseEntity.ok(response);
     }
